backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
django==5.0.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==4.4.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.27.2
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
psycopg==3.2.1
    # via -r requirements.in
psycopg-binary==3.2.1
    # via -r requirements.in
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.19.1
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via psycopg
tzdata==2024.1
    # via
    #   django
    #   psycopg
uritemplate==4.1.1
    # via drf-spectacular
